requirements.txt
customtkinter
routeros-api
keyring
pystray
Pillow